requirements.txt
# Deep Learning
torch==2.7.1
torchvision==0.20.1
tensorboard==2.18.0

# Web Framework
Flask==3.1.0
Flask-CORS==5.0.0

# Data Processing
numpy==1.24.3
Pillow==10.4.0

# Visualization
matplotlib==3.7.1

# Utilities
requests==2.31.0 